Обновила навигацию и создала структуру папок проекта
Добавила в main.dart логику выбора, какую стартовую страницу показывать в зависимости от того, пройдена ли регистрация и заполнена ли анкета.
Создала папочки в проекте по типу экранов и для данных.
Вынесла из main.dart экран логина в отдельный файл.
Вынесла из main.dart экран "Привет, друг!" в отдельный файл, теперь это экран списка помощников, там уже должен отображаться нижний аппбар
