+++ pubspec.yaml
@@ -55,8 +55,7 @@ dev_dependencies:
 
 # The following section is specific to Flutter packages.
 flutter:
-  assets:
-    - lib/telephone_code_verification/
+
 
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
